Fix vggt/pytorch: use random weights in compile-only mode and fix bfloat16 dtype mismatch
- Load model with random weights when TT_COMPILE_ONLY_SYSTEM_DESC is set
  to avoid gated repo access requirement
- Cast positional embeddings to input dtype in DPTHead._apply_pos_embed
  to prevent float32/bfloat16 mismatch caused by make_sincos_pos_embed

+++ vggt/pytorch/src/heads/dpt_head.py
@@ -312,7 +312,7 @@ class DPTHead(nn.Module):
         pos_embed = create_uv_grid(
             patch_w, patch_h, aspect_ratio=W / H, dtype=x.dtype, device=x.device
         )
-        pos_embed = position_grid_to_embed(pos_embed, x.shape[1])
+        pos_embed = position_grid_to_embed(pos_embed, x.shape[1]).to(dtype=x.dtype)
         pos_embed = pos_embed * ratio
         pos_embed = pos_embed.permute(2, 0, 1)[None].expand(x.shape[0], -1, -1, -1)
         return x + pos_embed
